pyproject.toml
[project]
name = "college-rl-rag-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "g4f>=0.6.0.2",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-cerebras>=0.5.0",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.1.9",
    "langchain-groq>=0.3.7",
    "langchain-nvidia-ai-endpoints>=0.3.16",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
